refactor: organize services in folders
