+++ src/app/animated-background-meme/animated-background-meme.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AnimationService } from '../services/animation.service';
+import { AnimationService } from '../services/animation/animation.service';
 
 import { animation } from './animations';
