+++ src/app/controls/controls.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { AnimationService } from '../services/animation.service';
-import { SudokuService } from '../services/sudoku.service';
+import { AnimationService } from '../services/animation/animation.service';
+import { SudokuService } from '../services/sudoku/sudoku.service';
 
 @Component({
   selector: 'app-controls',
