+++ src/app/services/animation/animation-assets.ts
